% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_endpoint}
\alias{add_endpoint}
\title{Add endpoint path to base URL}
\usage{
add_endpoint(path)
}
\arguments{
\item{path}{character. Endpoint path}
}
\value{
None.
}
\description{
Add endpoint path to base URL
}
