% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{get_categories}
\alias{get_categories}
\title{Get categories name and id}
\usage{
get_categories(api_key = Sys.getenv("ZOMATO_API_KEY"))
}
\arguments{
\item{api_key}{String. Api key.}
}
\value{
A tibble
}
\description{
A function to get a list of categories.
}
\examples{
\dontrun{
get_categories()
}
}
