% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{get_cities}
\alias{get_cities}
\title{Get city details}
\usage{
get_cities(
  q = NULL,
  lat = NULL,
  lon = NULL,
  city_ids = NULL,
  max = NULL,
  api_key = Sys.getenv("ZOMATO_API_KEY")
)
}
\arguments{
\item{q}{String. City name}

\item{lat}{Double. Latitude}

\item{lon}{Double. Longitude}

\item{city_ids}{Integer. City ID}

\item{max}{Integer. Number of max results to display}

\item{api_key}{String. Api key.}
}
\value{
A tibble
}
\description{
Find the Zomato ID and other details for a city.
}
\note{
You can obtain the Zomato City ID in one of the following ways: 1. City Name in the q (query) parameters - Returns list of cities matching the query, 2. Using coordinates (lat and lon) - Identifies the city details based on the coordinates of any location inside a city
}
\examples{
\dontrun{
get_cities(q = "new york", api_key = Sys.getenv("ZOMATO_API_KEY"))
}
}
