% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{get_collections}
\alias{get_collections}
\title{Get Zomato collections in a city}
\usage{
get_collections(
  city_id = NULL,
  lat = NULL,
  lon = NULL,
  max = NULL,
  api_key = Sys.getenv("ZOMATO_API_KEY")
)
}
\arguments{
\item{city_id}{Integer. City ID.}

\item{lat}{Double. Latitude.}

\item{lon}{Double. Longitude.}

\item{max}{Integer. Number of max results to display.}

\item{api_key}{String. Api key.}
}
\value{
A tibble
}
\description{
Returns Zomato restaurant collections in a city.
}
\note{
The location/City input can be provided in the following ways - 1. Using Zomato City ID; 2. Using coordinates of any location within a city.
}
\examples{
\dontrun{
# find 20 restaurant collections in New York city
get_collections(city_id = 61, api_key = Sys.getenv("ZOMATO_API_KEY"), max = 20)
}
}
