% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{get_cuisines}
\alias{get_cuisines}
\title{Get list of all cuisines in a city}
\usage{
get_cuisines(
  city_id = NULL,
  lat = NULL,
  lon = NULL,
  api_key = Sys.getenv("ZOMATO_API_KEY")
)
}
\arguments{
\item{city_id}{Integer. City ID}

\item{lat}{Double. Latitude}

\item{lon}{Double. Longitude}

\item{api_key}{String. Api key.}
}
\value{
A tibble
}
\description{
Get a list of all cuisines of restaurants listed in a city.
}
\note{
The location/city input can be provided in the following ways - 1. Using Zomato City ID, 2. Using coordinates of any location within a city.
}
\examples{
\dontrun{
# get list of all cuisines in New York city
get_cuisines(city_id = 61, api_key = Sys.getenv("ZOMATO_API_KEY"))
}
}
