% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restaurant.R
\name{get_dailymenu}
\alias{get_dailymenu}
\title{Get daily menu of a restaurant}
\usage{
get_dailymenu(res_id = NULL, api_key = Sys.getenv("ZOMATO_API_KEY"))

get_dailymenu(res_id = NULL, api_key = Sys.getenv("ZOMATO_API_KEY"))
}
\arguments{
\item{res_id}{Integer. ID of restaurant whose details are requested}

\item{api_key}{String. Api key.}
}
\value{
A tibble

A tibble
}
\description{
Get daily menu using Zomato restaurant ID.

Get detailed restaurant information using Zomato restaurant ID. Partner Access is required to access photos and reviews.
}
\examples{
\dontrun{
get_dailymenu(res_id = 16782899, api_key = Sys.getenv("ZOMATO_API_KEY"))
}
\dontrun{
get_restaurant(res_id = 16782899, api_key = Sys.getenv("ZOMATO_API_KEY"))
}
}
