% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_data}
\alias{get_data}
\title{Get the data from API}
\usage{
get_data(api_key = NULL, path, ...)
}
\arguments{
\item{api_key}{Character. User API key}

\item{path}{Character. Endpoint path.}

\item{...}{Further arguments passed to}
}
\description{
A function for requesting and parsing data from the API
}
