% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{get_establishments}
\alias{get_establishments}
\title{Get list of restaurant types in a city}
\usage{
get_establishments(
  city_id = NULL,
  lat = NULL,
  lon = NULL,
  api_key = Sys.getenv("ZOMATO_API_KEY")
)
}
\arguments{
\item{city_id}{Integer. City ID}

\item{lat}{Double. Latitude}

\item{lon}{Double. Longitude}

\item{api_key}{String. Api key.}
}
\value{
A tibble
}
\description{
Get a list of restaurant types in a city.
}
\note{
The location/City input can be provided in the following ways - 1. Using Zomato City ID, 2. Using coordinates of any location within a city.
}
\examples{
\dontrun{
# get list of restaurant types in New York city
get_establishments(city_id = 61, api_key = Sys.getenv("ZOMATO_API_KEY"))
}
}
