% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{get_geocode}
\alias{get_geocode}
\title{Get location details based on coordinates}
\usage{
get_geocode(lat = NULL, lon = NULL, api_key = Sys.getenv("ZOMATO_API_KEY"))
}
\arguments{
\item{lat}{Double. Latitude}

\item{lon}{Double. Longitude}

\item{api_key}{String. Api key.}
}
\value{
A list contains 3 tibble
}
\description{
Get Foodie and Nightlife Index, list of popular cuisines and nearby restaurants around the given coordinates.
}
\note{
The location/City input can be provided in the following ways - 1. Using Zomato City ID, 2. Using coordinates of any location within a city.
}
\examples{
\dontrun{
get_geocode(lat = 40.742051, lon = -74.004821,
}
}
