% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location.R
\name{get_location_details}
\alias{get_location_details}
\title{Get Zomato location details}
\usage{
get_location_details(
  entity_id = NULL,
  entity_type = NULL,
  api_key = Sys.getenv("ZOMATO_API_KEY")
)
}
\arguments{
\item{entity_id}{Integer. location id obtained from \code{get_locations}.}

\item{entity_type}{Character. location type obtained from \code{get_locations}.}

\item{api_key}{Character. Api key.}
}
\value{
A list
}
\description{
Get Foodie Index, Nightlife Index, Top Cuisines and Best rated restaurants in a given location.
}
\examples{
\dontrun{
get_location_details(entity_id = 114178, entity_type = subzone,
                    api_key = Sys.getenv("ZOMATO_API_KEY"))
}
}
