% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location.R
\name{get_locations}
\alias{get_locations}
\title{Search for locations}
\usage{
get_locations(
  query = NULL,
  lat = NULL,
  lon = NULL,
  max = NULL,
  api_key = Sys.getenv("ZOMATO_API_KEY")
)
}
\arguments{
\item{query}{String. suggestion for location name.}

\item{lat}{Double. Latitude}

\item{lon}{Double. Longitude}

\item{max}{Integer. Number of max results to display}

\item{api_key}{Character. Api key.}
}
\value{
A tibble
}
\description{
Search for Zomato locations by keyword. Provide coordinates to get better search results.
}
\examples{
\dontrun{
get_locations(query   = "london",
             api_key = Sys.getenv("ZOMATO_API_KEY"))
}
}
