% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restaurant.R
\name{get_restaurant}
\alias{get_restaurant}
\title{Get restaurant details}
\usage{
get_restaurant(res_id = NULL, api_key = Sys.getenv("ZOMATO_API_KEY"))
}
\arguments{
\item{res_id}{Integer. ID of restaurant whose details are requested}

\item{api_key}{String. Api key.}
}
\value{
A tibble
}
\description{
Get detailed restaurant information using Zomato restaurant ID. Partner Access is required to access photos and reviews.
}
\examples{
\dontrun{
get_restaurant(res_id = 16782899, api_key = Sys.getenv("ZOMATO_API_KEY"))
}
}
