% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restaurant.R
\name{get_reviews}
\alias{get_reviews}
\title{Get restaurant reviews}
\usage{
get_reviews(
  res_id = NULL,
  start = NULL,
  max = NULL,
  api_key = Sys.getenv("ZOMATO_API_KEY")
)
}
\arguments{
\item{res_id}{Integer. ID of restaurant whose details are requested}

\item{start}{Integer. Fetch results after this offset}

\item{max}{Integer. Max number of results to retrieve}

\item{api_key}{String. Api key.}
}
\value{
A tibble
}
\description{
Get restaurant reviews using the Zomato restaurant ID. Only 5 latest reviews are available under the Basic API plan.
}
\examples{
\dontrun{
get_reviews(res_id = 16782899, api_key = Sys.getenv("ZOMATO_API_KEY"))
}
}
