% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restaurant.R
\name{get_search}
\alias{get_search}
\title{Search for restaurants}
\usage{
get_search(
  q = NULL,
  entity_id = NULL,
  entity_type = NULL,
  lat = NULL,
  lon = NULL,
  radius = NULL,
  cuisines = NULL,
  establishment_type = NULL,
  collection_id = NULL,
  category = NULL,
  sort = NULL,
  order = NULL,
  is_free = TRUE,
  api_key = Sys.getenv("ZOMATO_API_KEY")
)
}
\arguments{
\item{q}{String. Search keyword.}

\item{entity_id}{Integer. Location id.}

\item{entity_type}{String. Location type. Possibel values: "city", "subzone", "landmark", "metro", or "group".}

\item{lat}{Double. Latitude.}

\item{lon}{Double. Longitude.}

\item{radius}{Double. Radius around (lat,lon); to define search area, defined in meters(M)}

\item{cuisines}{String. List of cuisine id's separated by comma.}

\item{establishment_type}{String. Establishment id obtained from \code{get_establishments}.}

\item{collection_id}{String. Collection id obtained from \code{get_collections}.}

\item{category}{String. Category ids obtained from \code{get_categories}.}

\item{sort}{String. sort restaurants by ... Possible values: "cost", "rating", or "real_distance".}

\item{order}{String. Used with 'sort' parameter to define ascending / descending. Possibel values: "asc", or "desc".}

\item{is_free}{Logical. If using free API set the value to TRUE}

\item{api_key}{String. Api key.}
}
\value{
A tibble
}
\description{
The location input can be specified using Zomato location ID or coordinates.
}
\note{
Cuisine / Establishment / Collection IDs can be obtained from respective api calls. Get up to 100 restaurants by changing the 'start' and 'count' parameters with the maximum value of count being 20. Partner Access is required to access photos and reviews.
}
\examples{
\dontrun{
# search for 'Italian' restaurants in 'Manhattan, New York City'
get_search(cuisines = 55,
          entity_id = 94741,
          entity_type = "zone",
          api_key = Sys.getenv("ZOMATO_API_KEY"))



          entity_type = "city",
          collection_id = 1)
}
}
