% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{handle_error}
\alias{handle_error}
\title{Error handler function}
\usage{
handle_error(x, msg)
}
\arguments{
\item{x}{list. List generated from API request}

\item{msg}{Character. The message to be displayed if the request is unsuccessful}
}
\description{
Functions to handle errors if the api request is unsuccessful.
}
