% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_key_null}
\alias{is_key_null}
\title{Api keys checker function}
\usage{
is_key_null(x = NULL)
}
\arguments{
\item{x}{character. API key}
}
\description{
Function to check whether the API key has been set.
}
