% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_api_key.R
\name{set_api_key}
\alias{set_api_key}
\title{Add API key to the .Renviron}
\usage{
set_api_key(path = stop("Please specify a path."))
}
\arguments{
\item{path}{character. Path where the environment is stored. Default is the normalized path.}
}
\value{
None.
}
\description{
Function to set you API Key to the R environment when starting using \code{zomator} package. Attention: You should only execute this functions once.
}
\examples{
\dontrun{
set_api_key(path = "/my/path/to/.Renviron")
}
}
