% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getcategories.R
\name{get_categories}
\alias{get_categories}
\title{category types}
\usage{
get_categories(key = NULL)
}
\arguments{
\item{key}{your API key}
}
\value{
category_id (integer) ID of the category type
category_name (string) Name of the category type
}
\description{
return the list of category types
}
\examples{
get_categories("testkey")
}
\author{
Yi Tang <tangyi2014@gmail.com>
}
