% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getcities.R
\name{get_cities}
\alias{get_cities}
\title{get cities ID}
\usage{
get_cities(key = NULL, q = NULL, lat = NULL, lon = NULL,
  city_ids = NULL, count = NULL)
}
\arguments{
\item{q}{query by city name}

\item{lat}{latitude}

\item{lon}{longitude}

\item{city_ids}{comma separated city_id values}

\item{count}{number of max results to display}

\item{user-key}{your API key}
}
\value{
id (integer) ID of the city
name (string) City name
country_id (integer, optional) ID of the country
country_name (string, optional) Name of the country
is_state (boolean, optional) Whether this location is a state
state_id (integer, optional) ID of the state
state_name (string, optional) Name of the state
state_code (string, optional) Short code for the state
}
\description{
search for Zomato ID and other details for a city
}
\examples{
get_cities(key="xxxxxx",q="xxx")
}
\author{
Yi Tang <tangyi2014@gmail.com>
}
