% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getcuisines.R
\name{get_cuisines}
\alias{get_cuisines}
\title{get cuisines ID and name}
\usage{
get_cuisines(key = NULL, city_id = NULL, lat = NULL, lon = NULL)
}
\arguments{
\item{city_id}{id of the city for which collections are needed}

\item{lat}{latitude}

\item{lon}{longitude}

\item{user-key}{your API key}
}
\value{
cuisine_id (integer) ID of the cuisine
cuisine_name (string) Name of the cuisine
}
\description{
search for all cuisines of restaurants listed in a city
}
\examples{
get_cuisines(key="xxxxxx",city_id=000)
}
\author{
Yi Tang <tangyi2014@gmail.com>
}
