% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdailymenu.R
\name{get_dailymenu}
\alias{get_dailymenu}
\title{Get daily menu of a restaurant}
\usage{
get_dailymenu(api_key = NULL, res_id = NULL)
}
\arguments{
\item{res_id}{id of restaurant whose details are requested}

\item{user-key}{your API key}
}
\value{
daily_menu_id (character) ID of the daily menu
start_date (character) start date of the daily menu
name (character) name of the daily menu
dishes (data frame) dishes on the daily menu
}
\description{
Get daily menu of a restaurant
}
\examples{
get_dailymenu(api_key="528b6ee8d624e5e3e741f1fbd895b760", res_id="16507624")
}
\author{
Chenxi Yang <chenxxiyang@gmail.com>
}
