% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getestablishments.R
\name{get_establishments}
\alias{get_establishments}
\title{get establishments}
\usage{
get_establishments(key = NULL, city_id = NULL, lat = NULL,
  lon = NULL)
}
\arguments{
\item{key}{API key, string}

\item{city_id}{id of the city, integer}

\item{lat}{latitude, double}

\item{lon}{longitude, double}
}
\value{
dataframe
establishment_id (integer) ID of the establishment type
establishment_name (string) Name of the establishment type
}
\description{
:
Get a list of restaurant types in a city.
}
\examples{
get_establishments(key = "testkey",city_id = 280)
}
\author{
Huazi Li <632473742@qq.com>
}
