% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getestablishments.R
\name{get_establishments}
\alias{get_establishments}
\title{get establishments}
\usage{
get_establishments(use_key, city_id, lat = NULL, lon = NULL)
}
\arguments{
\item{use_key}{API key, string}

\item{city_id}{id of the city, integer}

\item{lat}{latitude, double}

\item{lon}{longitude, double}
}
\value{
list
}
\description{
:
Get a list of restaurant types in a city. The
location/City input can be provided in the following ways:
1. Using Zomato City ID
2. Using coordinates of any location within a city
List of all restaurants categorized under a particular
restaurant type can obtained using /Search API with
Establishment ID and location details as inputs
}
\references{
\url{https://developers.zomato.com/documentation#!/common/establishments}
}
\author{
Huazi Li <632473742@qq.com>
}
