% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getgeocode.R
\name{get_geocode}
\alias{get_geocode}
\title{get geocode}
\arguments{
\item{use_key}{API key, string}

\item{lat}{latitude, double}

\item{lon}{longitude, double}
}
\value{
list
}
\description{
Get Foodie and Nightlife Index, list of popular
cuisines and nearby restaurants around the given coordinates
}
\references{
\url{https://developers.zomato.com/documentation#!/common/geocode}
}
\author{
Huazi Li <632473742@qq.com>
}
