% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getlocation_details.R
\name{get_location_details}
\alias{get_location_details}
\title{get location details}
\usage{
get_location_details(use_key, entity_id, entity_type)
}
\arguments{
\item{use_key}{API key, string}

\item{entity_id}{location id obtained from locations api, integer}

\item{entity_type}{location type obtained from locations api, string}
}
\value{
list
}
\description{
Get Foodie Index, Nightlife Index, Top Cuisines
and Best rated restaurants in a given location
}
\references{
\url{https://developers.zomato.com/documentation#!/location/location_details}
}
\author{
Huazi Li <632473742@qq.com>
}
