% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getlocations.R
\name{get_locations}
\alias{get_locations}
\title{get locations}
\usage{
get_locations(use_key, location_name, lat = NULL, lon = NULL,
  count = NULL)
}
\arguments{
\item{use_key}{API key, string}

\item{location_name}{suggestion for location name, string}

\item{lat}{latitude, double}

\item{lon}{longitude, double}

\item{count}{max number of results to fetch, integer}
}
\value{
list
}
\description{
:
Search for Zomato locations by keyword. Provide
coordinates to get better search results
}
\references{
\url{https://developers.zomato.com/documentation#!/location/locations}
}
\author{
Huazi Li <632473742@qq.com>
}
