% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getlocations.R
\name{get_locations}
\alias{get_locations}
\title{get locations}
\usage{
get_locations(key, query, lat = NULL, lon = NULL, count = NULL)
}
\arguments{
\item{key}{API key, string}

\item{query}{suggestion for location name, string}

\item{lat}{latitude, double}

\item{lon}{longitude, double}

\item{count}{max number of results to fetch, integer}
}
\description{
:
Search for Zomato locations by keyword. Provide
coordinates to get better search results
}
\examples{
get_locations("528b6ee8d624e5e3e741f1fbd895b760","van")

}
\author{
Huazi Li <632473742@qq.com>
}
