% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getrestaurant.R
\name{get_restaurant}
\alias{get_restaurant}
\title{Get restaurant details}
\usage{
get_restaurant(api_key = NULL, res_id = NULL)
}
\arguments{
\item{res_id}{id of restaurant whose details are requested}

\item{user-key}{your API key}
}
\description{
Get restaurant details
}
\examples{
get_restaurant(api_key="xxxxxx", res_id="xxxxxx")
}
\author{
Chenxi Yang <chenxxiyang@gmail.com>
}
