% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getrestaurant.R
\name{get_restaurant}
\alias{get_restaurant}
\title{Get restaurant details}
\usage{
get_restaurant(key = NULL, res_id = NULL)
}
\arguments{
\item{key}{your API key}

\item{res_id}{id of restaurant whose details are requested}
}
\description{
Get restaurant details
}
\examples{
get_restaurant(key="testkey", res_id="16774318")
}
\author{
Chenxi Yang <chenxxiyang@gmail.com>
}
