% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getreviews.R
\name{get_reviews}
\alias{get_reviews}
\title{Get restaurant reviews}
\usage{
get_reviews(key = NULL, res_id = NULL)
}
\arguments{
\item{key}{your API key}

\item{res_id}{id of restaurant whose details are requested}
}
\value{
rating (double) rating of the restaurant
review_text (character) review text of the restaurant
id (integer) id of the review provider
rating_color (character) rating_color of the review
review_time_friendly (character) time that the review is given
rating_text (character) rating_text about the review
timestamp (integer) timestamp of the review
likes (integer) likes of the review
comments_count (integer) number of comments of the review
user.name (character) user name of the review provider
user.foodie_level (character) user foodie_level of the review provider
user.foodie_level_num (integer) user foodie_level_num of the review provider
user.foodie_color (character) user foodie_color of the review provider
user.profile_url (character) url of the profile of the review provider
user.profile_image (character) image of the profile of the review provider
user.profile_deeplink (character) deeplink of the profile of the review provider
user.zomato_handle (character) zomato_handle about the user
}
\description{
Get restaurant reviews
}
\examples{
get_reviews(key="testkey", res_id="16774318")
}
\author{
Chenxi Yang <chenxxiyang@gmail.com>
}
