% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getsearch.R
\name{get_search}
\alias{get_search}
\title{Search for restaurants}
\usage{
get_search(api_key = NULL, q = NULL, entity_id = NULL,
  entity_type = NULL, start = NULL, count = NULL, lat = NULL,
  lon = NULL, radius = NULL, cuisines = NULL,
  establishment_type = NULL, collection_id = NULL, category = NULL,
  sort = NULL, order = NULL)
}
\arguments{
\item{entity_id}{location id}

\item{entity_type}{location type city/subzone/zone/landmark/metro/group}

\item{lat}{latitude}

\item{lon}{longitude}

\item{radius}{radius around (lat,lon)}

\item{cuisines}{list of cuisine id's separated by comma}

\item{establishment_type}{estblishment id obtained from establishments call}

\item{collection_id}{collection id obtained from collections call}

\item{sort}{sort restaurants by cost/rating/real_distance}

\item{order}{used with 'sort' parameter to define ascending/descending}

\item{user-key}{your API key}

\item{query}{search keyword}
}
\description{
Search for restaurants
}
\examples{
get_restaurant(api_key="xxxxxx", q="xxxxxx", entity_id="xxxxxx", entity_type="xxxxxx",
start="xxxxxx", count="xxxxxx", lat="xxxxxx", lon="xxxxxx",radius="xxxxxx",
cuisines="xxxxxx", establishment_type="xxxxxx", collection_id="xxxxxx",
category="xxxxxx", sort="xxxxxx"L, order="xxxxxx")
}
\author{
Chenxi Yang <chenxxiyang@gmail.com>
}
